pyproject.toml
[tool.poetry]
name = "gcp-mc"
version = "0.2.0"
description = "A minecraft server google cloud platform infrastructure project"
authors = ["Vasil Derilov"]
license = "GPL-3"
readme = "README.md"
packages = [{include = "gcp_mc"},
    {include = "utils", from="lib" }, 
    {include = "minecraft", from = "modules"}]

[tool.poetry.dependencies]
python = "^3.11"
google-api-python-client = "^2.78.0"
oauth2client = "^4.1.3"
docker = "^6.0.1"
pick = "^2.2.0"
pyyaml = "^6.0"
requests = "^2.28.2"
tqdm = "^4.64.1"


[tool.poetry.group.dev.dependencies]
pytest = "^7.2.1"
pylint = "^2.16.2"
flake8 = "^6.0.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.scripts]
gcp_mc_cli = 'gcp_mc:main'

[tool.black]
line-length = 88